+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file.
 
 - `libheif_info` function: added `encoders` and `decoders` keys to the result, for future libheif plugins support. #189
 - `options.PREFERRED_ENCODER` - to use `encoder` different from the default one. #192
+- `options.PREFERRED_DECODER` - to use `decoder` different from the default one. #193
 
 ### Changed